# calendly-clone-api/docs/en/api.yaml
openapi: 3.0.0
info:
  title: Calendly Clone API
  description: API for scheduling and managing appointments (English documentation)
  version: 1.0.0
servers:
  - url: /
    description: API endpoints
paths:
  /users:
    get:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: List all users
      description: Retrieve a paginated list of all users. Normal users will only see their own details, while admins can see all users.
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
          description: The page number to retrieve.
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of users to return per page.
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  - id: "user-123e4567-e89b-12d3-a456-426614174000"
                    name: "John Doe"
                    email: "john.doe@example.com"
                    timezone: "America/New_York"
                  - id: "user-223e4567-e89b-12d3-a456-426614174001"
                    name: "Jane Smith"
                    email: "jane.smith@example.com"
                    timezone: "Europe/London"
                  - id: "user-323e4567-e89b-12d3-a456-426614174002"
                    name: "Bob Johnson"
                    email: "bob.johnson@example.com"
                    timezone: "Asia/Tokyo"
                pagination:
                  page: 1
                  pageSize: 20
                  total: 3
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid pagination parameters"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user with a name, email, password, and optional timezone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              name: "John Doe"
              email: "john.doe@example.com"
              password: "securePassword123!"
              timezone: "America/New_York"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user-123e4567-e89b-12d3-a456-426614174000"
                name: "John Doe"
                email: "john.doe@example.com"
                timezone: "America/New_York"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c2VyLTEyM2U0NTY3LWU4OWItMTJkMy1hNDU2LTQyNjYxNDE3NDAwMCIsImlhdCI6MTUxNjIzOTAyMn0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email is already in use or invalid format"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Get user details
      description: Retrieve details of a specific user by their ID. Users can only access their own details unless they have admin privileges.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to retrieve.
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user-123e4567-e89b-12d3-a456-426614174000"
                name: "John Doe"
                email: "john.doe@example.com"
                timezone: "America/New_York"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "You can only access your own user details"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "User with ID user-123e4567 not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Partially update a user
      description: Update specific fields of a user by their ID. Users can only update their own information.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
            example:
              name: "John D. Doe"
              timezone: "Europe/Paris"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user-123e4567-e89b-12d3-a456-426614174000"
                name: "John D. Doe"
                email: "john.doe@example.com"
                timezone: "Europe/Paris"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid timezone format"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Forbidden: You can only access or modify your own data"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "User with ID user-123e4567 not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: []
      tags:
        - Users
      summary: Delete a user
      description: Delete a specific user by their ID. Users can only delete their own account.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user to delete.
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Forbidden: You can only access or modify your own data"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions:
    post:
      tags:
        - Authentication
      summary: Create session (login)
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
            example:
              email: "john.doe@example.com"
              password: "securePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c2VyLTEyM2U0NTY3LWU4OWItMTJkMy1hNDU2LTQyNjYxNDE3NDAwMCIsImlhdCI6MTUxNjIzOTAyMn0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                user:
                  id: "user-123e4567-e89b-12d3-a456-426614174000"
                  name: "John Doe"
                  email: "john.doe@example.com"
                  timezone: "America/New_York"
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email and password are required"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid email or password"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "An error occurred during authentication"

    delete:
      tags:
        - Authentication
      summary: Destroy session (logout)
      description: End user session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Successfully logged out"
        '400':
          description: No token provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "No token provided"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "An error occurred during logout"
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
      required:
        - page
        - pageSize
        - total
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        timezone:
          type: string
        token:
          type: string  # Token is optional
      required:
        - id
        - name
        - email
        - password
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad Request"
      required:
        - message
    UserPatch:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        timezone:
          type: string
      minProperties: 1  # At least one field must be provided
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token